refactor(collab): move the expressId bound helpers to lib/collab
From a local CodeRabbit review. Taking the narrow half and declining the wide
half, on the record.

TAKEN: the two pure helpers move out of the 1300-line slice into
`lib/collab/express-id-bounds.ts`, with their test alongside. The reason is not
the line count: #2706 moves the reconstruct itself into
`lib/collab/room-model-apply.ts`, so the rule now sits where that refactor will
want it rather than having to be carried across in the same change. Pure
functions are also testable without driving an async reconstruct nothing can
start, which is what made the original defect untestable in place.

DECLINED: 'continue splitting until the production module is near the ~400 line
guideline'. collabSlice.ts was ~1200 lines before this PR touched it; splitting
it is a real piece of work with its own review surface, and doing it inside a
bug fix would bury the fix. The size rule is self-policed here with no gate,
which is a reason to schedule the split deliberately, not to smuggle it in.

Re-verified after the move, and the move itself nearly hid that: the mutation
script still pointed at the OLD path, so all four mutations silently applied to
nothing and reported no failures - which reads exactly like a passing run. With
the path corrected, all four are caught by named tests again, subtest count
unchanged at 10.
