+++ apps/viewer/src/store/slices/collabSlice.ts
@@ -73,6 +73,7 @@ import {
   seedFailureMessage,
   writeGeometrySeedMarker,
 } from '@/lib/collab/geometry-seed-signal';
+import { highestExpressId, raisedMaxExpressId } from '@/lib/collab/express-id-bounds';
 import { createSharedBlobStore } from '@/lib/collab/blob-store';
 import {
   attachAnnotationInbound,
@@ -90,41 +91,6 @@ import { getEntityCenter } from '@/utils/viewportUtils';
  * UI affordances.
  */
 export type CollabRole = 'viewer' | 'commenter' | 'editor' | 'admin';
-
-/**
- * Highest expressId in a reconstructed id map, or 0 when there is none.
- *
- * One definition for both reconstruct branches: the first build and every
- * re-derive have to agree on what the bound means, and computing it twice by
- * hand is how they came to disagree (#2719).
- */
-export function highestExpressId(idToPath: Map<number, string> | undefined): number {
-  let max = 0;
-  if (idToPath) {
-    for (const id of idToPath.keys()) if (id > max) max = id;
-  }
-  return max;
-}
-
-/**
- * The new `maxExpressId` bound for a re-derived room model, or `null` when it
- * must not move.
- *
- * RAISED, NEVER LOWERED. A peer deleting an entity shrinks the id space, but
- * ids already handed out elsewhere (selection, annotations, a pending mesh)
- * have to keep resolving through `globalId.ts`'s
- * `localExpressId <= model.maxExpressId` gate, so the bound is a high-water
- * mark rather than a current count. Lowering it would turn a delete into a
- * silent unresolvable-selection bug, which is the same defect as the freeze,
- * just triggered from the other direction.
- */
-export function raisedMaxExpressId(
-  current: number,
-  idToPath: Map<number, string> | undefined,
-): number | null {
-  const seen = highestExpressId(idToPath);
-  return seen > current ? seen : null;
-}
 
 export type CollabStatus = 'disconnected' | WebSocketStatus | 'memory' | 'indexeddb';
